Fix trick for `make BINDDIR=. shell` and `COPY .`
This was inadvertently removed when updating the Dockerfile for buildkit
specific features.

Trick selects a different build target depending on if the source is
going to be bind-mounted in anyway, which prevents the need to copy the
whole source tree to the builder.

Signed-off-by: Brian Goff <cpuguy83@gmail.com>
(cherry picked from commit 1c82d1190888f05836c550c85e4e8daedae2a813)