Enable step-by-step debugging
This basically forces the use of a newer Go version -- one recent enough
to allow Delve (dlv) to run on the container in headless mode. We can
then connect to it through port 40000 from the host OS and see the code
alive. Great for debugging 🐛 and learning about the code base 🔬.