distribution: add back resumable pulls
Upstream docker supports resuming pulls after transient failures using
HTTP range requests. In commit 565fc34, this was replaced with streaming
the tarball directly to disk, obviating the need for a temporary file,
and reducing the amount of free space required to download images.

It also aimed to reduce the writes to flash memory. In practice, this
tends to cause more writes when large images fail to download, and
the download is forced to restart. On unstable connections with very
large images, the download never completes.

Add back the upstream method with resumable pulls in place of the
streaming approach.

Change-type: patch
Signed-off-by: Joseph Kogut <joseph@balena.io>